requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.9.0
pandas>=2.2.0
python-dotenv>=1.0.1
anthropic>=0.36.0
Flask>=3.0.0
